moved searchUsers function to usersState & create Search component for HTML, also DataTable using now UsersState for loading users & for search result
